ui: reverts to using base forms to avoid infinite rendering bug
Initially we would have liked form-components to basically build themselves
if a form wasn't passed in. Right now we don't need this, and we found
that the approach we used produced an 'infinite rendering' bug but
_only_ during actual use (we've tried to write a test to replicate it
but it doesn't manifest during acceptance testing).

We've rethought how this works, but as we don't need to do it yet we've
reverted to retrieving forms on the base component