ui: Test (scenario steps) changes to allow access to deeper properties
We always has the 'click $property on the $component component'-like
steps

This was never going to be sufficient longer term, and with the form
within form layouts we needed something so we can traverse deeper into
page objects.

This add javascript like traversing for some of the scenario steps (to
be expanded upon at a later date) so you can use dot syntax like
'roles.form.policies.form.selectedOptions.0'

Note this also adds the ability to 'select' indexes on collections using
'.0' or '.1' etc. But we are keeping the default to the first element if
you don't specify an index as it reads nicver.

Lastly, this amends the tests that required this